requirements.txt
Flask==2.3.2
PyGithub==1.77
radon==5.2.0
pyflakes==3.0.0
jinja2==3.1.2
